Schafe laufen nun zwischen mehreren Kacheln, ernten eine gewisse Zeit und produzieren anschließend Wolle/Schlachtvieh - TODOs: aktuell sind noch viele Sachen fix im Code, die per Config gesteuert werden müssen - TODO: später soll die Produktion von Wolle/Schlachtvieh nicht direkt bei Rückkehr des Trägers, sondern durch Umwandlung des unsichtbaren Input-Guts gemacht werden
Carrier
- Carrier::initRoute() enthält nun mehr gemeinsamen Code
- haben nun Status und Info, wie lange sie zum Ernten brauchen

Tests
- haben nun die Möglichkeit, ein MapObjectType zu verändern
- Helper-Makro ASSERT_NO_ENDLESS_LOOP, um Endlosschleifen zu vermeiden
- Test hinzugefügt, der korrektes Verhalten der Schafe überprüft
- Test für InCatchmentAreaFinder hinzugefügt, der Route zwischen Map-Kacheln testet
- Test checkThatCarriersAreSentCorrectly aktiviert

Sonstiges
- InCatchmentAreaFinder::findMapTileWithInvisibleGood() kann nun auch Routen zwischen zwei Map-Kacheln finden (d.h. das Zentralgebäude ist nicht Startpunkt der Route)
- GraphicSet kann nun direkt zu EightDirectionsAnimation umgewandelt werden
- Fix: Schaf-Grafiken werden nun korrekt eingelesen
- TICKS_PER_*-Konstanten sind nun Ganzzahlen
- Nomenklator: grassland -> grass

+++ src/cpp/economics/EconomicsMgr.cpp
@@ -34,25 +34,25 @@ void EconomicsMgr::updateProduction(Building* building) {
     // Haben wir Eingabegüter, dann wird nur produziert, wie diese verfügbar sind
     if (mapObjectType->buildingProduction.input.isUsed()) {
         inputConsumed =
-            (double) ticksSinceLastUpdate / TICKS_PER_MINUTE * mapObjectType->buildingProduction.input.rate;
+            (double) ticksSinceLastUpdate / (double) TICKS_PER_MINUTE * mapObjectType->buildingProduction.input.rate;
 
         // nur verbrauchen, was auch da is
         if (inputConsumed > building->productionSlots.input.inventory) {
             inputConsumed = building->productionSlots.input.inventory;
         }
         ticksInputConsumed =
-            (unsigned int) (inputConsumed * TICKS_PER_MINUTE / mapObjectType->buildingProduction.input.rate);
+            (unsigned int) (inputConsumed * (double) TICKS_PER_MINUTE / mapObjectType->buildingProduction.input.rate);
 
         if (mapObjectType->buildingProduction.input2.isUsed()) {
             input2Consumed =
-                (double) ticksSinceLastUpdate / TICKS_PER_MINUTE * mapObjectType->buildingProduction.input2.rate;
+                (double) ticksSinceLastUpdate / (double) TICKS_PER_MINUTE * mapObjectType->buildingProduction.input2.rate;
 
             // nur verbrauchen, was auch da is
             if (input2Consumed > building->productionSlots.input2.inventory) {
                 input2Consumed = building->productionSlots.input2.inventory;
             }
             ticksInput2Consumed =
-                (unsigned int) (input2Consumed * TICKS_PER_MINUTE / mapObjectType->buildingProduction.input2.rate);
+                (unsigned int) (input2Consumed * (double) TICKS_PER_MINUTE / mapObjectType->buildingProduction.input2.rate);
         }
     }
 
@@ -68,15 +68,15 @@ void EconomicsMgr::updateProduction(Building* building) {
 
     // Jetzt die Produktion durchführen
     if (mapObjectType->buildingProduction.input.isUsed()) {
-        inputConsumed = (double) ticksWeReallyProduced / TICKS_PER_MINUTE * mapObjectType->buildingProduction.input.rate;
+        inputConsumed = (double) ticksWeReallyProduced / (double) TICKS_PER_MINUTE * mapObjectType->buildingProduction.input.rate;
         building->productionSlots.input.decreaseInventory(inputConsumed);
 
         if (mapObjectType->buildingProduction.input2.isUsed()) {
             input2Consumed =
-                (double) ticksWeReallyProduced / TICKS_PER_MINUTE * mapObjectType->buildingProduction.input2.rate;
+                (double) ticksWeReallyProduced / (double) TICKS_PER_MINUTE * mapObjectType->buildingProduction.input2.rate;
             building->productionSlots.input2.decreaseInventory(input2Consumed);
         }
     }
-    outputProduced = (double) ticksWeReallyProduced / TICKS_PER_MINUTE * mapObjectType->buildingProduction.output.rate;
+    outputProduced = (double) ticksWeReallyProduced / (double) TICKS_PER_MINUTE * mapObjectType->buildingProduction.output.rate;
     building->productionSlots.output.increaseInventory(outputProduced);
 }
